refactor: remove inactive duplicate shared sources
