+++ tests/.last_test_status.json
@@ -1,11 +1,11 @@
 {
   "files": {
     "tests/test_cpp_unit.py": {
-      "last_run": "2026-03-30T11:07:40",
+      "last_run": "2026-03-30T14:23:22",
       "status": "PASS"
     },
     "tests/test_interface.py": {
-      "last_run": "2026-03-30T11:07:40",
+      "last_run": "2026-03-30T14:23:22",
       "status": "PASS"
     },
     "tests/test_optional_testdata.py": {
@@ -13,7 +13,7 @@
       "status": "PASS"
     },
     "tests/test_st_gt.py": {
-      "last_run": "2026-03-30T11:07:40",
+      "last_run": "2026-03-30T14:23:22",
       "status": "PASS"
     },
     "tests/test_unit_alignment.py": {
@@ -38,9 +38,9 @@
     }
   },
   "session": {
-    "command": "PHYLOACC_RUN_GT=1 PHYLOACC_RUN_TESTDATA=1 /n/home07/gthomas/miniconda3/envs/phyloacc-test/bin/python -m pytest -q",
+    "command": "PHYLOACC_RUN_GT=1 /n/home07/gthomas/miniconda3/envs/phyloacc-test/bin/python -m pytest tests/test_interface.py tests/test_st_gt.py tests/test_cpp_unit.py -q",
     "exitstatus": 0,
-    "summary": "33 passed in 234.59s",
-    "timestamp": "2026-03-30T11:07:40"
+    "summary": "13 passed in 206.44s",
+    "timestamp": "2026-03-30T14:23:22"
   }
 }
